+++ cellprofiler_core/image/abstract_image_provider/load_images_image_provider/_load_images_image_provider.py
@@ -8,7 +8,8 @@ import six.moves
 from cellprofiler_core.image._image import Image
 import cellprofiler_core.preferences
 from .._abstract_image_provider import AbstractImageProvider
-from ....utilities.image import is_numpy_file, is_matlab_file, loadmat, load_data_file, pathname2url, url2pathname
+from ....utilities.image import is_numpy_file, is_matlab_file, loadmat, load_data_file
+from ....utilities.pathname import pathname2url, url2pathname
 
 from cellprofiler_core.utilities import generate_presigned_url
