+++ cellprofiler_core/pipeline/_pipeline.py
@@ -1999,7 +1999,7 @@ class Pipeline:
         """
         if isinstance(path_or_fd, six.string_types):
             from cellprofiler_core.modules import FILE_SCHEME
-            from cellprofiler_core.utilities.image import url2pathname
+            from cellprofiler_core.utilities.pathname import url2pathname
 
             pathname = path_or_fd
 
@@ -2034,7 +2034,7 @@ class Pipeline:
 
     def add_pathnames_to_file_list(self, pathnames, add_undo=True):
         """Add a sequence of paths or URLs to the file list"""
-        from cellprofiler_core.utilities.image import pathname2url
+        from cellprofiler_core.utilities.pathname import pathname2url
 
         urls = []
         for pathname in pathnames:
