+++ cellprofiler_core/utilities/image.py
@@ -247,54 +247,3 @@ def load_data_file(pathname_or_url, load_fn):
 
 def is_file_url(url):
     return url.lower().startswith(FILE_SCHEME)
-
-
-def urlfilename(url):
-    """Return just the file part of a URL
-
-    For instance http://nucleus.org/linked_files/file%20has%20spaces.txt
-    has a file part of "file has spaces.txt"
-    """
-    if is_file_url(url):
-        return os.path.split(url2pathname(url))[1]
-    path = six.moves.urllib.parse.urlparse(url)[2]
-    if "/" in path:
-        return six.moves.urllib.unquote(path.rsplit("/", 1)[1])
-    else:
-        return six.moves.urllib.unquote(path)
-
-
-def urlpathname(url):
-    """Return the path part of a URL
-
-    For instance, http://nucleus.org/Comma%2Cseparated/foo.txt
-    has a path of http://nucleus.org/Comma,separated
-
-    A file url has the normal sort of path that you'd expect.
-    """
-    if is_file_url(url):
-        return os.path.split(url2pathname(url))[0]
-    scheme, netloc, path = six.moves.urllib.parse.urlparse(url)[:3]
-    path = six.moves.urllib.parse.urlunparse([scheme, netloc, path, "", "", ""])
-    if "/" in path:
-        return six.moves.urllib.unquote(path.rsplit("/", 1)[0])
-    else:
-        return six.moves.urllib.unquote(path)
-
-
-def pathname2url(path):
-    """Convert the unicode path to a file: url"""
-    utf8_path = six.text_type(path)
-    if any([utf8_path.lower().startswith(x) for x in PASSTHROUGH_SCHEMES]):
-        return utf8_path
-    return FILE_SCHEME + six.moves.urllib.request.pathname2url(utf8_path)
-
-
-def url2pathname(url):
-    if isinstance(url, six.text_type):
-        url = url
-    if any([url.lower().startswith(x) for x in PASSTHROUGH_SCHEMES]):
-        return url
-    assert is_file_url(url)
-    utf8_url = six.moves.urllib.request.url2pathname(url[len(FILE_SCHEME) :])
-    return utf8_url
