+++ tests/pipeline/test_pipeline.py
@@ -1169,12 +1169,12 @@ HasImagePlaneDetails:False"""
         assert len(p.file_list) == 2
         for path in paths:
             assert (
-                    cellprofiler_core.utilities.image.pathname2url(path) in p.file_list
+                    cellprofiler_core.utilities.pathname.pathname2url(path) in p.file_list
             )
 
     def test_read_file_list_urls(self):
         root = os.path.split(__file__)[0]
-        file_url = cellprofiler_core.utilities.image.pathname2url(
+        file_url = cellprofiler_core.utilities.pathname.pathname2url(
             os.path.join(root, "foo.tif")
         )
         urls = [
